pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "soc-analyst-env"
version = "1.0.0"
description = "Cybersecurity Threat Response Agent — OpenEnv environment for investigating and responding to security alerts"
requires-python = ">=3.10"
dependencies = [
    "openenv-core>=0.1.0",
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "pydantic>=2.0.0",
]

[project.scripts]
server = "server.app:main"

[project.optional-dependencies]
inference = [
    "openai>=1.0.0",
    "requests>=2.28.0",
]
